Pkg cleanup: moving out crd from pkg. Adjusting all dependencies
