+++ crd/apis/danm/register.go
