+++ crd/apis/danm/v1/doc.go
@@ -3,4 +3,3 @@
 // Package v1 is the v1 version of the API.
 // +groupName=danm.k8s.io
 package v1
-
