+++ pkg/ipam/ipam.go
@@ -10,8 +10,8 @@ import (
   "time"
   "math/big"
   "math/rand"
-  danmtypes "github.com/nokia/danm/pkg/crd/apis/danm/v1"
-  danmclientset "github.com/nokia/danm/pkg/crd/client/clientset/versioned"
+  danmtypes "github.com/nokia/danm/crd/apis/danm/v1"
+  danmclientset "github.com/nokia/danm/crd/client/clientset/versioned"
   meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
   "github.com/nokia/danm/pkg/danmnet"
   "github.com/nokia/danm/pkg/bitarray"
